Preparations for upgrading PDFBox from 1.8.16 to 2.0.22.

+++ test/src/org/lockss/pdf/MockPdfTokenFactory.java
@@ -110,7 +110,7 @@ public class MockPdfTokenFactory implements PdfTokenFactory {
   @Override
   public PdfToken makeObject(final PdfToken value,
                              long objectNumber,
-                             long generationNumber) {
+                             int generationNumber) {
     return new MockPdfToken() {
       @Override public PdfToken getObject() { return value; }
       @Override public boolean isObject() { return true; }
